Rename and GCC fixes
