+++ src/fwd/fwd_profile.h
@@ -1,4 +1,4 @@
-#include "abs.h"
+#include "fwd_abs.h"
 
 namespace fwd::profile {
 struct full {
